requirements.txt
fastapi == 0.104.1
uvicorn[standar] == 0.24.0
sqlalchemy == 2.0.23
psycopg2-binary == 2.9.9
python-dotenv == 1.0.0
pydantic[email] == 2.5.0
alembic == 1.12.1